Refactor to reorder place classes
Refactor to a DDD structure
