+++ server/src/common/place/place.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { DbModule } from '../db/db.module';
-import { PlaceRepository } from './place.repository';
-import { PlaceController } from './place.controller';
+import { PlaceRepository } from './persistence/place.repository';
+import { PlaceController } from './rest/place.controller';
 
 @Module({
     modules: [DbModule],
